Move all stubs to same package
